requirements.txt
Django==1.11.7
wheel==0.30.0
gunicorn==19.7.1
psycopg2==2.7.3.2
nltk==3.2.5
djangorestframework==3.7.0
geopy==1.11.0
django-cors-headers==2.0.0


